feat: support for type = 'Plain' feat: support for properties 'Visited' and 'Current' in links. chore: updated styling of sitemap component. chore: renamed CSS class 'nested' to 'ct-nested'.
