+++ src/api/factory.js
@@ -1,4 +1,4 @@
-import createHtmlContentProperty from './html-content-property'
+import contentProperty from './content-property'
 import createObjectProperty from './object-property'
 import createPrimitiveProperty from './primitive-property'
 import createArrayProperty from './array-property'
@@ -7,8 +7,8 @@ export default function(data, parent) {
   if (data.type == 'object') {
     return createObjectProperty(data, parent)
   }
-  if (data.type == 'html') {
-    return createHtmlContentProperty(data, parent)
+  if (data.type == 'html' || data.type == 'plain') {
+    return contentProperty(data, parent)
   }
   if (data.type == 'array') {
     return createArrayProperty(data, parent)
