+++ src/components/elements/factory.js
@@ -3,6 +3,7 @@ import HtmlContent from './html-content'
 import Link from './link'
 import Section from './section'
 import List from './list'
+import Plain from './plain'
 
 export default property => {
   if (property.links.navigate) {
@@ -11,6 +12,7 @@ export default property => {
 
   switch (property.type) {
     case 'html': return HtmlContent
+    case 'plain': return Plain
     case 'object': return Section
     case 'array' : return List
     default: return Input
