+++ src/components/elements/html-content.js
@@ -1,6 +1,8 @@
-export default ({ property }) => (
+import cx from 'classnames'
+
+export default ({ className, property }) => (
   <div
-    className='ct-content'
+    className={cx(className, 'ct-content')}
     dangerouslySetInnerHTML={{ __html: property.content }}
     id={property.id} />
 )
