+++ src/components/elements/list.js
@@ -1,4 +1,5 @@
 import factory from './factory'
+import cx from 'classnames'
 
 export default ({ property }) => (
   <div className='ct-list'>
@@ -7,7 +8,9 @@ export default ({ property }) => (
 
       return (
         <Element
-          className='ct-list-item'
+          className={cx('ct-list-item', {
+            'ct-invalid' : item.errors && item.errors.length})}
+          errors={item.errors}
           key={item.id}
           property={item} />
       )
