+++ themes/default/app.css
@@ -6,6 +6,7 @@
 @import './json-debugger.css';
 @import './link.css';
 @import './list.css';
+@import './plain.css';
 @import './section.css';
 @import './sitemap.css';
