+++ themes/default/link.css
@@ -1,10 +1,21 @@
 .ct-link {
-  color: #5DB0EA;
   font-size: 15px;
   font-style: normal;
   font-weight: 300;
   padding: 5px 0;
   text-decoration: none;
+}
+
+.ct-link, .ct-list-item.ct-link {
+  color: #5DB0EA;
+}
+
+.ct-link.ct-invalid {
+  color: #ED1631;
+}
+
+.ct-link.ct-invalid:hover {
+  color: #CC0302;
 }
 
 .ct-link:hover {
