+++ themes/default/plain.css
@@ -1,0 +1,3 @@
+.ct-plain {
+  margin: 0;
+}
